First Graph added!  Basic run distance over last 12 weeks

+++ FitnessViewer.Infrastructure/Helpers/MetreDistance.cs
@@ -11,9 +11,9 @@ namespace FitnessViewer.Infrastructure.Helpers
     /// </summary>
     public static class MetreDistance
     {
-        private const double METRE_TO_MILE = 0.00062137119;
-        private const double METER_TO_KM = 0.001;
-        private const double METER_TO_FEET = 3.2808399;
+        private const float METRE_TO_MILE = 0.00062137119F;
+        private const float METER_TO_KM = 0.001F;
+        private const float METER_TO_FEET = 3.2808399F;
         private const int PRECISION = 2;
 
         /// <summary>
@@ -21,9 +21,9 @@ namespace FitnessViewer.Infrastructure.Helpers
         /// </summary>
         /// <param name="metres">metre distance to convert</param>
         /// <returns></returns>
-        public static double ToMiles(double metres)
+        public static float ToMiles(float metres)
         {
-            return Math.Round(metres * METRE_TO_MILE, PRECISION);
+            return (float)Math.Round(metres * METRE_TO_MILE, PRECISION);
         }
 
         /// <summary>
@@ -31,9 +31,9 @@ namespace FitnessViewer.Infrastructure.Helpers
         /// </summary>
         /// <param name="metres">metre distance to convert</param>
         /// <returns></returns>
-        public static double ToKM(double metres)
+        public static float ToKM(float metres)
         {
-            return Math.Round(metres * METER_TO_KM, PRECISION);
+            return (float)Math.Round(metres * METER_TO_KM, PRECISION);
         }
 
         /// <summary>
@@ -41,9 +41,9 @@ namespace FitnessViewer.Infrastructure.Helpers
         /// </summary>
         /// <param name="metres">metre distance to convert</param>
         /// <returns></returns>
-        public static double ToFeet(double metres)
+        public static float ToFeet(float metres)
         {
-            return Math.Round(metres * METER_TO_FEET, PRECISION);
+            return (float)Math.Round(metres * METER_TO_FEET, PRECISION);
         }
     }
 }
